Rename package structure from nl.programit.rollingsticks back to nl.rollingsticks
