+++ .gitignore
@@ -4,3 +4,4 @@ target/*
 .project
 /target/
 /bin/**
+/src/main/resources/application.properties
