+++ src/main/java/nl/rollingsticks/RollingsticksApplication.java
@@ -1,4 +1,4 @@
-package nl.programit.rollingsticks;
+package nl.rollingsticks;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
