+++ src/main/java/nl/rollingsticks/domain/TekstService.java
@@ -1,4 +1,4 @@
-package nl.programit.rollingsticks.domain;
+package nl.rollingsticks.domain;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
