+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
 spring.datasource.url=jdbc:mysql://localhost:3306/rollingsticks
 spring.datasource.driverClassName=com.mysql.jdbc.Driver
-spring.datasource.username=rollingsticks
-spring.datasource.password=rollitaround12345
+spring.datasource.username=root
+spring.datasource.password=
 
 spring.jpa.database_platform=org.hibernate.dialect.MySQL5Dialect
 #NOT IN PRODUCTION. NICE IN DEVELOPMENT!!!
